third_party/blink/web_tests/fast/events/popup-blocking-timers1-expected.txt

Click Here
PASS successfullyParsed is true

TEST COMPLETE

Test calling window.open() directly. A popup should be allowed.
PASS newWindow is non-null.
